+++ ykrt/src/trace/mod.rs
@@ -106,6 +106,6 @@ impl TracedAOTBlock {
 }
 
 #[cfg(tracer_swt)]
-pub(crate) fn trace_basicblock(function_index: u32, block_index: u32) {
+pub(crate) fn trace_basicblock(function_index: usize, block_index: usize) {
     swt::trace_basicblock(function_index, block_index)
 }
